Create Homework Assignment 1
Write a program that plays a work game with the user. The program should ask the user to enter the following.
His or her name
His or her age
The name of a city
The name of a college
A profession
A type of animal
A pet's name

After the user has entered these items, the program should display the following story
inserting the user's input into the appropriate locations:

"There once was a person name NAME who lived in CITY. At the age of AGE, Name went to college at COLLEGE.
NAME graduated and went to work as a PROFESSION. Then, NAME adopted a(n) ANIMAL name PETNAME. They both lived happily ever after!